navigation/requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
customtkinter==5.2.2
darkdetect==0.8.0
decorator==5.1.1
future==1.0.0
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.1
idna==3.10
iso8601==2.1.0
packaging==24.2
pillow==11.0.0
pyperclip==1.9.0
pyserial==3.5
PyYAML==6.0.2
ratelim==0.1.6
requests==2.32.3
six==1.16.0
tkintermapview==1.29
urllib3==2.2.3
